fix validate lương

+++ src/main/java/com/DATN/DTO/responseDTO/SalaryCalculationResponseDTO.java
@@ -13,7 +13,7 @@ public class SalaryCalculationResponseDTO {
 
     private Integer workingDays;
 
-    private Double basicSalary;
+    private Integer basicSalary;
 
     private Double allowance;
